backend-python/requirements.txt
# FastAPI & Web Framework
fastapi==0.111.0
uvicorn[standard]==0.30.0
python-multipart==0.0.9
pydantic==2.8.0
pydantic-settings==2.3.0

# Database
sqlalchemy==2.0.29
psycopg[binary]>=3.2.0
alembic==1.13.1

# AI/ML Libraries
torch==2.9.1
torchvision==0.24.1
torchaudio==2.9.1
transformers==4.40.0
diffusers==0.27.0
accelerate==0.30.0
opencv-python==4.9.0.80
pillow>=10.4.0
numpy>=1.26.0

# MediaPipe
mediapipe>=0.10.13

# Celery & Redis
celery==5.3.6
redis==5.0.6

# Cloud Storage
boto3==1.34.90

# Security
passlib[bcrypt]==1.7.4
python-jose[cryptography]==3.3.0

# Cloud AI
google-genai>=0.1.0

# Utilities
PyPDF2==3.0.1
pandas==2.2.2
openpyxl==3.1.2
python-dotenv==1.0.1
httpx==0.27.0
aiofiles==23.2.1
